refactor: break each step in AssignmentMetadataModal into subcomponent
